fix : 알림 관련 매직 넘버는 설정파일로 관리한다.
