+++ src/test/java/life/offonoff/ab/application/notice/NoticeServiceIntegrationTest.java
@@ -5,16 +5,16 @@ import life.offonoff.ab.domain.TestEntityUtil;
 import life.offonoff.ab.domain.member.Member;
 import life.offonoff.ab.domain.topic.Topic;
 import life.offonoff.ab.web.response.notice.NoticeResponse;
-import life.offonoff.ab.web.response.notice.VoteCountOnTopicNoticeResponse;
+import life.offonoff.ab.web.response.notice.message.VoteCountOnTopicNoticeMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static life.offonoff.ab.application.notice.NoticeService.VOTE_COUNT_MODULO;
 import static life.offonoff.ab.domain.TestEntityUtil.createRandomMember;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -25,9 +25,28 @@ public class NoticeServiceIntegrationTest {
 
     @Autowired
     NoticeService noticeService;
+    @Value("${ab.notification.vote_on_topic.count_unit}")
+    int voteCountUnit;
 
     @Autowired
     EntityManager em;
+
+
+    @Test
+    @DisplayName("Vote Count 알림 기준에 해당하면 알림을 생성한다.")
+    void topic_voteCount_should_noticed() {
+        // given
+        Topic topic = TestEntityUtil.TestTopic.builder()
+                .id(1L)
+                .voteCount(voteCountUnit)
+                .build().buildTopic();
+
+        // when
+        boolean shouldNotice = noticeService.shouldNoticeVoteCountForTopic(topic);
+
+        // then
+        assertThat(shouldNotice).isTrue();
+    }
 
     @Test
     @DisplayName("VoteCountOnTopic 알림 생성한다.")
@@ -38,7 +57,7 @@ public class NoticeServiceIntegrationTest {
 
         Topic topic = TestEntityUtil.TestTopic.builder()
                 .author(author)
-                .voteCount(VOTE_COUNT_MODULO)
+                .voteCount(voteCountUnit)
                 .build().buildTopic();
         em.persist(topic);
 
@@ -48,6 +67,6 @@ public class NoticeServiceIntegrationTest {
         List<NoticeResponse> responses = noticeService.findAllByReceiverId(author.getId());
 
         // then
-        assertThat(responses.get(0)).isInstanceOf(VoteCountOnTopicNoticeResponse.class);
+        assertThat(responses.get(0).getMessage()).isInstanceOf(VoteCountOnTopicNoticeMessage.class);
     }
 }
