+++ src/test/java/life/offonoff/ab/application/notice/NoticeServiceTest.java
@@ -2,14 +2,10 @@ package life.offonoff.ab.application.notice;
 
 import life.offonoff.ab.domain.member.Member;
 import life.offonoff.ab.domain.notice.DefaultNotification;
-import life.offonoff.ab.domain.notice.Notification;
-import life.offonoff.ab.domain.notice.VoteCountOnTopicNotification;
-import life.offonoff.ab.domain.notice.VoteResultNotification;
 import life.offonoff.ab.domain.topic.Topic;
 import life.offonoff.ab.domain.vote.VoteResult;
 import life.offonoff.ab.repository.member.MemberRepository;
 import life.offonoff.ab.repository.notice.NotificationRepository;
-import life.offonoff.ab.web.response.notice.DefaultNoticeResponse;
 import life.offonoff.ab.web.response.notice.NoticeResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -18,11 +14,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static life.offonoff.ab.application.notice.NoticeService.VOTE_COUNT_MODULO;
 import static life.offonoff.ab.domain.TestEntityUtil.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -88,22 +82,6 @@ class NoticeServiceTest {
 
         // then
         assertThat(author.getNotifications().size()).isGreaterThan(0);
-    }
-
-    @Test
-    @DisplayName("Vote Count 알림 기준에 해당하면 알림을 생성한다.")
-    void topic_voteCount_should_noticed() {
-        // given
-        Topic topic = TestTopic.builder()
-                .id(1L)
-                .voteCount(VOTE_COUNT_MODULO)
-                .build().buildTopic();
-
-        // when
-        boolean shouldNotice = noticeService.shouldNoticeVoteCountForTopic(topic);
-
-        // then
-        assertThat(shouldNotice).isTrue();
     }
 
     @Test
